Fix a bug
When an equation was formed and = was clicked and a result
was displayed, then a number was pressed, what happened
was that the number got appended to the equation.

Example in steps:
================

equation displayed: 2 + 5     |    result displayed: 7
A digit such as 2 is inserted
equation now: 2 + 52          | result displayed: 7

I have fixed that. Now, when the equation is full and a result
is displayed, and a digit is inserted, the data is cleared
and the calculator starts over by replacing the equation
with the newly inserted input.